pyproject.toml
[tool.poetry]
name = "b2broker"
version = "0.1.0"
description = ""
authors = ["Anatoly <anatolyfeteleu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
django = "^5.0.4"
django-rest-framework = "^0.1.0"
django-environ = "^0.11.2"
mysqlclient = "^2.2.4"
ipython = "^8.23.0"
django-filter = "^24.2"
factory-boy = "^3.3.0"
pytest-django = "^4.8.0"
ipdb = "^0.13.13"

[tool.black]
line-length = 79
